1.5.7.14-dev
You can now run '/session start finale' to activate a finale session. This session will be infinite and will have any finale specific features.
Wild Life finale - callback will be activated + wildcard nerfs.
Secret Life finale - players are give the 'Win Secret Life' task.

+++ src/main/java/net/mat0u5/lifeseries/seasons/season/wildlife/wildcards/wildcard/Callback.java
@@ -59,7 +59,7 @@ public class Callback extends Wildcard {
         int passedTimeTicks = currentSession.getPassedTime().getTicks();
         int sessionLengthTicks = currentSession.isInfiniteSession() ? infiniteSessionFakeTime.getTicks() : currentSession.getSessionLength().getTicks();
 
-        double sessionProgress = (passedTimeTicks -activatedAt) / (sessionLengthTicks -activatedAt);
+        double sessionProgress = (double) (passedTimeTicks - activatedAt) / (sessionLengthTicks - activatedAt);
 
         if (nextActivationTick == -1) {
             nextActivationTick = passedTimeTicks + INITIAL_ACTIVATION_INTERVAL; // First activation after 5 minutes
